Replace fallback function with fund() definition
Let's keep the fallback function simple, in this case let's remove it.

Fallback functions are called when a contract is sent a message with no
arguments (or when no function matches), and only has access to 2,300
gas when called from a `.send()` or `.transfer()`.

See https://github.com/ConsenSys/smart-contract-best-practices/blob/master/README.md#keep-fallback-functions-simple