Replace original contract
The new contract will work for all the campaigns. Instead of deploying
a new instance of the contract for every new campaign, the new
contract will be deployed once and it will keep a mapping between
issue ids and _campaigns_. Each campaign is an instance of the
`Campaign` struct, that includes more or less all the attributes that
were present in the old contract.

There are multiple reasons why a single contract should be better than
multiple contracts. First of all it should be cheaper in terms of gas
to create a new struct value instead of deploying a new contract
instance. A possible workaround was to store most of the functions in
a library and to delegate function calls. Another reason why it seems
better to deploy a single contract is because there will be a single
entry point to call contract functions.