Test rollbackFunds function
The function gives to funders the ability to cancel their funding in
the case a resolution has been verified, but they're not happy with
the solution. They have up to `defaultRewardPeriod` seconds from the
moment the resolution has been verified, that is one day, to call
this function to get back their Eth.

A fairer (future) version of this function, after a funder rollbacks
his/her fund, should keep part of the amount in the campaign. This
will prevent cases when the author of the resolution could have
worked for nothing, because all the funders call this function right
after a proposed resolution has been verified, draining all of the
campaign's funds.