Add tests for withdrawFunds
`withdrawFunds` is used by the author of the accepted solution to
withdraw a campaign's funds as a reword for his work. This is the
best function ever! I really hope people will enjoy using it.

There are a couple of main checks:

* a campaign can't be rewarded multiple times
* a campaign must be rewarded only to the address that submitted a
  solution that has been verified
* campaign's funds can be withdrawn only after a grace period of
  one day
* campaign's funds can't be withdrawn after 7 days from the end of
  the grace period

Note: I'd like to remove the last condition. I added a TODO note
about it, let's work on it later.