Update verifyResolution
This update fixes a potential security issue. The function will throw
an exception if the resolutor and its associated patch commit SHA
don't match to the one that has been passed as a function argument.
This will prevent concurrent updates of a patch submitted by the same
author.

This special case could happen when a patch's author submits two
different commits in sequence and, while the issue manager is sending
a transaction to verify the first patch, the second submission
transaction is executed.

See https://github.com/issuehunter/issuehunter/pull/1#discussion_r135197823