Update submitPatch to handle verification fee
The new `submitPatch` is a _payable_ method that accepts a
verification fee. The sender must pay a patch verification fee that
will be used by the patch verifier, after the patch has been
verified, to inform the contract of the successful verification.

The patch verification fee amount is set to twice the amount of Ether
needed to send a transaction to execute the method `verifyPatch`
according the gas price in input. In theory this should be more than
enough for the verifier to run the transaction and to spare some gas.

See https://github.com/issuehunter/issuehunter/issues/13.