Store tip per mille value
A default tip value that is applied if a custom tip value hasn't been
specified during campaign creation. The default tip is 5%.

`createCampaignWithVerifier` has been renamed to
`createCampaignExtended` and it includes both a custom patch verifier
address and a custom tip value.

Tip values are expressed in 1000th.

A valid tip value range is between 10 (1%) and 200 (20%).

In case of invalid tip value the contract throws an exception.