Calculate tips
Tips are "applied" when a patch is verified. The tips amount is stored
in a new `tipsAmount` campaign's attribute and it's added to the
`tipsAmount` contract's attribute.

Campaign's attribute is used to keep historical track of campaigns
tips.

Contract's attribute will be used later for the implementation of a
method that can be used by the contract's owner to withdraw tips.
After each tips withdrawal the attribute should be reset to zero.

Funders and the verified patch author will be able to withdraw their
Ether minus the application of a campaign's tip per mille, stored in
a campaign's `tipsPerMille` attribute.

** A note about tips per mille computation **

Withdrawable amounts are calculated as reciprocal of a `tipsPerMille`
value. All amounts are `uint` and in Solidity integer division
truncates the result. For this reason the total amount of
withdrawable, that is the sum of reciprocals, should always be less
than the reciprocal of the total campaign amount. This means that the
sum of withdrawable amounts + the campaign's `tipsAmount` is less
than or equal to the total withdrawable amount + the campaign's
`tipsAmount`.

Example:

funds[backerA] = 10
funds[backerB] = 10
funds[backerC] = 10
tipsPerMille = 50

totalAmount = funds.sum = 10 + 10 + 10 = 30
withdrawableReward
  = totalAmount * (1000 - tipsPerMille) / 1000
  = 30 * (1000 - 50) / 1000
  = 28.5
  = 28 (int)
tipsAmount
  = totalAmount - withdrawableReward
  = 30 - 28
  = 2

withdrawableAmounts[backerA]
  = funds[backerA] * (1000 - tipsPerMille) / 1000
  = 10 * (1000 - 50) / 1000
  = 9.5
  = 9 (int)
withdrawableAmounts[backerB]
  = funds[backerB] * (1000 - tipsPerMille) / 1000
  = 10 * (1000 - 50) / 1000
  = 9.5
  = 9 (int)
withdrawableAmounts[backerC]
  = funds[backerC] * (1000 - tipsPerMille) / 1000
  = 10 * (1000 - 50) / 1000
  = 9.5
  = 9 (int)

In this example if all funders rollback their funds after the patch
has been verified, they will be able to withdraw 27, that is less than
or equal to the total withdrawable amount, that is 28, but the
contract owner will be able to withdraw only 2 and the contract will
hold a spare amount of 1.